requirements.txt
torch
torchvision
fastapi
uvicorn
pillow
scikit-learn
pdf2image